Add functionality to save extracted text
This negates the need to re-extract pdf text every time the terms file changes, as long as the input PDF dataset stays the same.

However, nested dicts in the txt file aren't currently being parsed properly (garbled output) so this change will be rolled back until it can be implemented.